requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
requests==2.32.3
schedule==1.2.2
urllib3==2.2.3
